+++ src/sql/changes/update_2021-11-01_conform_widgets.sql
@@ -1,0 +1,9 @@
+
+-- Update stats widgets
+UPDATE project_widgets
+SET widget = jsonb_build_object(
+    'name', widget->>'name',
+    'type', 'statistics',
+    'layout', widget->'layout'
+)
+WHERE widget->'properties'->>0 = 'getStats';
